requirements.txt
ffmpeg
discord.py[voice] ~= 1.4.0
PyNaCl==1.4.0
youtube-dl==2021.4.26
youtube-search-python==1.4.7
